Apply pr feedback.
